Correct deserialize of ObjectSpaceObjectPropSet
According to [MS-ONESTORE], section 2.6.1, the optional field ContextIDs is
only present if OSIDs is present and the OSIDs.header.ExtendedStreamsPresent
field is true.

This commit nests the check whether there is a ContextIDs field present
into the branch where the optional field OSIDs is present.
If OSIDs is not present, ContextIDs cannot be present as well according
to the specification.